+++ src/main/java/kitchenpos/table/dto/OrderTableResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 import java.util.Objects;
