+++ src/main/java/kitchenpos/table/dto/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.dto;
+package kitchenpos.table.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
