+++ src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
+import kitchenpos.table.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
