+++ src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
 import kitchenpos.common.fixtrue.OrderTableFixture;
+import kitchenpos.order.domain.OrderTables;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
